Organizing files in folders. Start playing with loans and FRED data.
